# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: LocationKit
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /location:
    x-swagger-router-controller: location
    get:
      operationId: getAllLocations
      description: get locations list
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetLocationsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: saveLocation
      description: add location to the list
      parameters:
        - name: location
          description: Location properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Location"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /location/{id}:
    x-swagger-router-controller: location
    get:
      operationId: getOneLocation
      description: get a location
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetLocationResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: updateLocation
      description: update a location
      parameters:
        - name: id
          description: location id
          type: string
          in: path
          required: true
        - name: location
          description: location data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Location"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delLocation
      description: delete a location
      parameters:
        - name: id
          description: location id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /collection:
    x-swagger-router-controller: collection
    get:
      operationId: getAllCollections
      description: add a collection
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetCollectionsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: saveCollection
      description: add a collection
      parameters:
        - name: collection
          description: collection that is being added
          in: body
          required: true
          schema:
            $ref: "#/definitions/Collection"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /collection/{id}:
    x-swagger-router-controller: collection
    get:
      operationId: getOneCollection
      description: get a single collection
      parameters:
        - name: id
          description: collection id
          in: path
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetCollectionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: updateCollection
      description: update a collection
      parameters:
        - name: id
          description: collection id
          type: string
          in: path
          required: true
        - name: collection
          description: collection to update
          in: body
          required: true
          schema:
            $ref: "#/definitions/Collection"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delCollection
      description: delete a collection
      parameters:
        - name: id
          description: collection id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  GetCollectionsListResponse:
    required:
      - collections
    properties:
      collections:
        type: array
        items:
          $ref: "#/definitions/Collection"
  Collection:
    required:
      - id
      - locations
    properties:
      id:
        type: string
      name:
        type: string
      locations:
        type: array
        items:
          $ref: "#/definitions/Location"
      info: 
        type: object
  GetCollectionResponse:
    required:
      - collection
    properties:
      collection:
        $ref: "#/definitions/Collection"
  GetLocationsListResponse:
    required:
      - locations
    properties:
      locations:
        type: array
        items:
          $ref: "#/definitions/Location"
  GetLocationResponse:
    required:
      - location
    properties:
      location:
        $ref: "#/definitions/Location"
  Location:
    required:
      - id
    properties:
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      street_address:
        type: string
      zip_code:
        type: string
      city_id:
        type: string
      country_id:
        type: string
      info:
        type: object
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string